[host] Move HAL bodies into components/host/hal.cpp + inline trivial dispatches
Per-platform follow-up to #15977.

Move the multi-line clock_gettime/nanosleep wrappers (millis, millis_64,
delay, micros, delayMicroseconds, arch_restart, arch_get_cpu_cycle_count)
from components/host/core.cpp into a new components/host/hal.cpp.
core.cpp keeps only the host bootstrap (signal handlers + main()).

Inline the trivial one-liners and empty bodies in hal_host.h:

  yield()                  -> ::sched_yield()
  arch_init()              -> {}
  arch_feed_wdt()          -> {}
  arch_get_cpu_freq_hz()   -> 1000000000U
